feat(cozeloop-ai): duration -> duration_micros
