+++ packages/cozeloop-ai/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cozeloop/ai",
-  "version": "0.0.4-beta.3",
+  "version": "0.0.4-beta.4",
   "description": "Official Node.js SDK of CozeLoop | 扣子罗盘官方 Node.js SDK",
   "keywords": [
     "cozeloop",
